The downloadAndCheckPdfFile and downloadAndCheckXlsFile tests are moved to the new DownloadAndCheckPdfTest class
